+++ projects/ngx-stripe/src/lib/services/stripe.service.ts
@@ -49,13 +49,10 @@ import {
   CreateTokenCardData,
   CreateTokenPiiData,
   CreateTokenBankAccountData,
-  PaymentIntent,
-  PaymentMethod,
   PaymentRequest,
   PaymentRequestOptions,
   RedirectToCheckoutOptions,
   RetrieveSourceParam,
-  SetupIntent,
   Stripe,
   StripeCardElement,
   StripeCardNumberElement,
@@ -66,8 +63,6 @@ import {
   StripeElement,
   StripeError,
   StripeIbanElement,
-  Source,
-  Token,
   TokenCreateParams,
   ConfirmBacsDebitSetupData,
   ConfirmBancontactSetupData,
@@ -112,7 +107,16 @@ import {
   CollectBankAccountTokenOptions,
   EphemeralKeyNonceOptions,
   StripeElementsOptionsClientSecret,
-  StripeElementsOptionsMode
+  StripeElementsOptionsMode,
+  ConfirmBlikPaymentData,
+  ConfirmBlikPaymentOptions,
+  ConfirmCashappPaymentData,
+  ConfirmCashappPaymentOptions,
+  PaymentIntentOrSetupIntentResult,
+  CreatePaymentMethodFromElements,
+  CreatePaymentMethodFromElement,
+  ConfirmCashappSetupData,
+  ConfirmCashappSetupOptions
 } from '@stripe/stripe-js';
 
 import { STRIPE_PUBLISHABLE_KEY, STRIPE_OPTIONS, NGX_STRIPE_VERSION } from '../interfaces/ngx-stripe.interface';
@@ -177,10 +181,22 @@ export class StripeService implements StripeServiceInterface {
     redirect: 'if_required';
   }): Observable<PaymentIntentResult>;
   confirmPayment(options: {
+    elements?: StripeElements;
+    clientSecret: string;
+    confirmParams?: Partial<ConfirmPaymentData>;
+    redirect: 'if_required';
+  }): Observable<PaymentIntentResult>;
+  confirmPayment(options: {
     elements: StripeElements;
     confirmParams: ConfirmPaymentData;
     redirect?: 'always';
   }): Observable<never | { error: StripeError }>;
+  confirmPayment(options: {
+    elements?: StripeElements;
+    clientSecret: string;
+    confirmParams: ConfirmPaymentData;
+    redirect?: 'always';
+  }): Observable<never | {error: StripeError}>;
   confirmPayment(options) {
     return this.stripe.confirmPayment(options);
   }
@@ -223,6 +239,14 @@ export class StripeService implements StripeServiceInterface {
     return this.stripe.confirmBancontactPayment(clientSecret, data, options);
   }
 
+  confirmBlikPayment(
+    clientSecret: string,
+    data?: ConfirmBlikPaymentData,
+    options?: ConfirmBlikPaymentOptions
+  ): Observable<PaymentIntentResult> {
+    return this.stripe.confirmBlikPayment(clientSecret, data, options);
+  }
+
   confirmBoletoPayment(
     clientSecret: string,
     data?: ConfirmBoletoPaymentData,
@@ -237,6 +261,14 @@ export class StripeService implements StripeServiceInterface {
     options?: ConfirmCardPaymentOptions
   ): Observable<PaymentIntentResult> {
     return this.stripe.confirmCardPayment(clientSecret, data, options);
+  }
+
+  confirmCashappPayment(
+    clientSecret: string,
+    data?: ConfirmCashappPaymentData,
+    options?: ConfirmCashappPaymentOptions
+  ): Observable<PaymentIntentResult> {
+    return this.stripe.confirmCashappPayment(clientSecret, data, options);
   }
 
   confirmCustomerBalancePayment(
@@ -371,6 +403,10 @@ export class StripeService implements StripeServiceInterface {
     return this.stripe.handleCardAction(clientSecret);
   }
 
+  handleNextAction(options: { clientSecret: string; }): Observable<PaymentIntentOrSetupIntentResult> {
+    return this.stripe.handleNextAction(options);
+  }
+
   verifyMicrodepositsForPayment(
     clientSecret: string,
     data?: VerifyMicrodepositsForPaymentData
@@ -386,8 +422,11 @@ export class StripeService implements StripeServiceInterface {
     return this.stripe.collectBankAccountForPayment(options);
   }
 
-  createPaymentMethod(paymentMethodData: CreatePaymentMethodData): Observable<PaymentMethodResult> {
-    return this.stripe.createPaymentMethod(paymentMethodData);
+  createPaymentMethod(paymentMethodData: CreatePaymentMethodData): Observable<PaymentMethodResult>;
+  createPaymentMethod(options: CreatePaymentMethodFromElements): Observable<PaymentMethodResult>;
+  createPaymentMethod(options: CreatePaymentMethodFromElement): Observable<PaymentMethodResult>;
+  createPaymentMethod(options) {
+    return this.stripe.createPaymentMethod(options);
   }
 
   retrievePaymentIntent(clientSecret: string): Observable<PaymentIntentResult> {
@@ -400,10 +439,22 @@ export class StripeService implements StripeServiceInterface {
     redirect: 'if_required';
   }): Observable<SetupIntentResult>;
   confirmSetup(options: {
+    elements?: StripeElements;
+    clientSecret: string;
+    confirmParams?: Partial<ConfirmPaymentData>;
+    redirect: 'if_required';
+  }): Observable<SetupIntentResult>;
+  confirmSetup(options: {
     elements: StripeElements;
     confirmParams: ConfirmPaymentData;
     redirect?: 'always';
   }): Observable<never | { error: StripeError }>;
+  confirmSetup(options: {
+    elements?: StripeElements;
+    clientSecret: string;
+    confirmParams: ConfirmPaymentData;
+    redirect?: 'always';
+  }): Observable<never | {error: StripeError}>;
   confirmSetup(options) {
     return this.stripe.confirmSetup(options);
   }
@@ -438,6 +489,14 @@ export class StripeService implements StripeServiceInterface {
     options?: ConfirmCardSetupOptions
   ): Observable<SetupIntentResult> {
     return this.stripe.confirmCardSetup(clientSecret, data, options);
+  }
+
+  confirmCashappSetup(
+    clientSecret: string,
+    data?: ConfirmCashappSetupData,
+    options?: ConfirmCashappSetupOptions
+  ): Observable<SetupIntentResult> {
+    return this.stripe.confirmCashappSetup(clientSecret, data, options);
   }
 
   confirmIdealSetup(clientSecret: string, data?: ConfirmIdealSetupData): Observable<SetupIntentResult> {